[Fixes #516] : Restructure timers code
